Return created attributes rather than errors
I am now making a big design decision that is slightly abstract from
writing Go: rather than returning errors, we're going to return
attributes, and if errors are detected, we're going to immediately fail.

This means we can use "outputs" of resources to work with other
resources, which can be extremely useful.

I do not think we should expect the user to have to wrap every resource
creation in an error check, since it will make the configuration a lot
more verbose.