Add some platform attributes
Since we're compiling for a specific platform, we can use the Go
compiler attributes to designate the OS and Arch.

If it's a Linux platform, then we can populate details about the system
using the `/etc/os-release` file.

If this file doesn't exist, then rather than error, we'll just return
empty values.