Add git resource
This adds a git resource which can clone repositories. I manage quite a
bit of my stuff using git repositories so it's quite useful.

At the moment all it does is clone a repository, and does nothing if it
already exists. However, we should be able to pass in a branch and
whether or not to fetch latest changes etc.