Add template helpers
In the end, using the `embed`[1] and `template`[2] packages made it
trivial to embed files and templates into our code.

I did think about adding a whole new resource called `template`, but in
the end it seemed overkill, because all we really want to do is generate
a string that writes to a file.

I think there may be some other uses of embedded files when we probably
shouldn't output them to a string, but for small files it seems like an
OK way to do it.

[1] https://pkg.go.dev/embed
[2] https://pkg.go.dev/text/template