Remove sudo support
I've decided I would prefer running the program using sudo, but having
the ability to set a default user.

Some resources enforce the use of using root.

This does complicate things, particularly in the user of how we might
authenticate with SSH keys.

As such, we need to add the following to `/etc/sudoers` to allow us to
authenticate with our current ssh agent:

```
Defaults env_keep+=SSH_AUTH_SOCK
```