Implement concurrent resource operations
This commit implements the v2 approach of viaduct. It works in this way:

Firstly, we create a map of resources at the start of every run. This is
done by creating a new defined map with `run := v.New()`.

We then add items to this map in order by generating a unique ID. The
unique ID is a JSON representation of the resources generated, which
means that if we're declaring exactly the same resource, then it will
throw an error. It's possible we may wish to declare the same resource
twice, and we could do that by adding some unique naming data, but I'm
not sure how common that would be.

These IDs can be used for dependency management, where dependencies can
be declared for each resources by the ID they return (see the example in
this commit).

Once we've created our map, we start the concurrent run by submitting
`run.Start()`.

There are still a few things left to handle:

1. The Package and Apt resources have *implicit* dependences, because
   generally they cannot run at the same time. We could handle this
   directory in viaduct, which would involve searching the map for
   another of those dependencies; or we could just use the package lock
   mechanisms they implement themselves, and wait for them to become
   available.

2. If we adopt this model, we should probably return an error when we
   run: `err := run.Start()`. However, my original model just
   immediately raised an error which simplified the code. To implement
   this, each resource would have to return errors to trickle back. I
   think this would be straight forward to add.

I also think there's a lot of boiler plate and it could be refactored,
but the basics are there.