Remove unneccesary verbs
I think we can get away with only using Create and Delete for verbs
which makes it easier to remember how to use different resources. The
exception here is the Execute resource, where Create and Delete does not
make sense. Possibly we could just "create" an execution but not include
it for "Delete".