Pass back errors rather than immediately quitting
This implements proper error handling with concurrency. Rather than just
immediately killing the process, we pass errors back, and handle the
errors from dependencies. Ideally we want to run everything, but if a
dependency throws an error, then anything that depends on it refuses to
run.

It also implements dumping out the manifest in the case of an error so
the user can inspect it. Using a hashed resource name is actually kinda
clumsy in this case, and I could probably calculate some contextual
resource names based upon the attributes.

I've implemented it so that scripting mode is unaffected.