Improve error handling
This improves error handling by including the attributes of the
resources with the errors at the end of the run.

I've updated the constants to be plainer since I didn't think it was
useful having eg OperationCreate rather than just Create printed.

Since we're now dumping errors I've turned the manifest dumping into an
optional flag.