Fiddling with new syntax
One thing that I would like to be able to achieve is the ability to keep
a consistent interface for plugging in new resources.

Think about how each resource is configured and then called, and see how
much logic we can cut out in the manifest.go and resource.go files, with the
logic put more into the resource type files instead.

One thing that could be improved is that we probably don't need to
specify Create or Delete; we can just have Delete as a bool that must be
explicitly set, and assume that we want to do the default action. This
ensures the logic stays within the resource type.