Move resources into separate package
The split between the resources and the logic required to run the
resources was getting a bit mixed up, so I think given that I intend to
allow any custom resource to be written, it makes sense to move
resources to their own separate package. The resources here are
considered "stdlib".