Ensure that preflight checks run before everything
We need to ensure that preflight checks happen before we start a run,
since there's no point waiting to discover that Viaduct configuration
for a resource is misconfigured.

I'm not bothering with concurrency since it should be very quick to
iterate over all the resources to perform very simple checks. If
preflights ever get a little more complicated, then concurrency should
be considered.