Add helper commands
Rather than including Unless statements in all the different resources,
we can actually just embed useful helpers right into the configuration.
This is the approach I'm taking with deciding whether to declare
resources. This does mean that the requirement of the resources are
evaluated before the run starts, so in some cases we may wish to still
evaluate these during the run itself.