Add file functionality to download resource
It should set permissions like other resources.

Ideally we need some helpers here and more standardisation to set
permissions and these default set of conditions that are useful
everywhere.

Perhaps a Meta type that can be applied to resources, and then some
methods on that to provide all the needed functionality.