Rename Config to Cli
I was thrown when I wanted to output attributes, since the flag was
`--output-attributes`, when I assumed it would just be `--atributes`.

Upon further investigation it seems that `Configs` really is not a good
name for what it's doing, because it only really contains CLI stuff, so
I have renamed it accordingly.